Sqlite hustle execution integrated with resolver (4/n) (#65)
* get the catalog information in the resolver based on db name

* Add arrow fields and schema as member for the Schema classes

* Catalog tests for the arrow fields in the hustle schema classes

* added query execution from resolver

* added scheduling capability in query execution

* added a query exectution with the resolver

* remove dead code - commented codes

* fix dependencies issue

* add license header

* change the specifer to PUBLIC in cmakelist

* remove unused resolver files

* removing resolver header file

+++ src/resolver/select_resolver.cc
@@ -1,3 +1,20 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 #include "resolver/select_resolver.h"
 
 #include <memory>
@@ -46,7 +63,6 @@ std::shared_ptr<PredicateTree> SelectResolver::ResolvePredExpr(Expr* pExpr) {
   if (pExpr == NULL) {
     return nullptr;
   }
-  // std::cout << "Operation " << (int)pExpr->op << std::endl;
   arrow::compute::CompareOperator comparatorOperator;
   FilterOperator connective;
   std::shared_ptr<PredicateTree> predicate_tree = nullptr;
@@ -103,7 +119,7 @@ std::shared_ptr<PredicateTree> SelectResolver::ResolvePredExpr(Expr* pExpr) {
           datum = arrow::Datum(
               std::make_shared<arrow::StringScalar>(rightExpr->u.zToken));
         } else if (rightExpr->op == TK_INTEGER) {
-          datum = arrow::Datum((uint32_t)rightExpr->u.iValue);
+          datum = arrow::Datum((int64_t)rightExpr->u.iValue);
         }
         if (pExpr->op == TK_NE)
           comparatorOperator = arrow::compute::CompareOperator::NOT_EQUAL;
